LPS-67048 Move check properties to base class
